# Optimize ObjectId parsing