--- a/GitCommands/Git/GitModule.cs
+++ b/GitCommands/Git/GitModule.cs
@@ -1163,12 +1163,12 @@ namespace GitCommands
             string submodule = lines[0];
 
             if (submodule.Length < ObjectId.Sha1CharCount + 3
-                || !ObjectId.TryParse(submodule, 1, out ObjectId objectId))
+                || !ObjectId.TryParse(submodule, 1, out ObjectId currentCommitId))
             {
                 return (' ', null);
             }
 
-            return (submodule[0], objectId);
+            return (submodule[0], currentCommitId);
         }
 
         public bool ExistsMergeCommit(string? startRev, string? endRev)