--- a/UnitTests/GitCommands.Tests/GitRevisionInfoProviderTests.cs
+++ b/UnitTests/GitCommands.Tests/GitRevisionInfoProviderTests.cs
@@ -32,9 +32,7 @@ namespace GitCommandsTests
             var item = Substitute.For<IGitItem>();
 
             // ObjectId checks input, use Try to get an illegal value
-            ObjectId objectId;
-            var sourceBytes = Encoding.ASCII.GetBytes("");
-            ObjectId.TryParseAsciiHexBytes(sourceBytes, 0, out objectId);
+            ObjectId.TryParse("", out ObjectId objectId);
             item.ObjectId.Returns(objectId);
 
             ((Action)(() => _provider.LoadChildren(item))).Should().Throw<ArgumentException>();