## Proposed changes

Use optimized Utf8Parser(Span<>) methods to parse Git commit hashes.

Remove unused methods.

Benchmark for `TryParse(in ReadOnlySpan<byte> array)` that run in average 3.1 times for each commit (for linux repo)
To parse 100000 commits (default in GE) was before 290 ms, now it is 110 ms
The total load time therefore reduced with 550 ms. For GE this is about 160 ms.

## Test methodology <!-- How did you ensure quality? -->

Tests are updatedNote that some test cases were removed when unused methods were removed.

## Merge strategy

I agree that the maintainer squash merge this PR (if the commit message is clear).

----

:black_nib: I contribute this code under [The Developer Certificate of Origin](../blob/master/contributors.txt).