@vbjay 2023-09-13
I see it is utf8 specific. Are we forcing utf8 by using i18n.logOutputEncoding and other related i18n.* config. https://git-scm.com/docs/git-config#Documentation/git-config.txt-i18nlogOutputEncoding

Git default and/or dedacto may be utf8 but are we sure every single one is set to utf8?  One of the many reasons I keep harping on  system git confif not being read. We need to at the least read config values from system and then global and so on.