@gerhardol 2023-09-13
> I see it is utf8 specific. Are we forcing utf8 

This is a potential issue. This PR has the same assumption as the current code, that 0-9,a-f is encoded using ASCII.
Using UTF8parser slightly improves the compatibility...

Similar apply to the timestamp with unix time stamp, also assumed to be ASCII 0-9
(This PR is really a spin-off from a change in RevisionReader, but other changes are needed there for UTF8parser to be faster.

For author names, email, subject etc the data is decoded with the encoding reported by Git, config is read.
(I am adding a comment about the potential issue).

utf-8 should probably be forced for git-log. There are some optimized encoders in .NET8 I believe.

For the hash, timestamps I propose to keep it as is though also if utf-8 is not forced, there will be a performance impact if all is decoded.
(uint.TryParse(Span<char>) is slower than UTF8Parser.TryParse(Span<byte for some reason too).